[feat/#35]: 아지트 소개 수정 구현

+++ src/main/java/org/crews/dto/request/IntroducingRequest.java
@@ -21,6 +21,6 @@ public class IntroducingRequest {
 
     private String content;
 
-    private List<Integer> deleteInterests;
-    private List<Integer> addInterests;
+    private List<Long> deleteInterests;
+    private List<Long> addInterests;
 }
